Re-order dynamic config tests to end of bucket, and swap to FULL mode
